Grind recipes and alchemy ingredients (#3600)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Removed the organ grind recipe that made viscera
leeches and worms are grindable
Removes the whole... pot-recipe-thing-system that was no longer being
used
any item when examined will tell you their smell, rather than just the
special alchemy items
Also made it so when grinding poppies you get a chance for extra ozium

## Why It's Good For The Game

So you can grind organs for their poison (why would you do that honestly
no idea but some wanted so here it is)
leeches are a cheap (but not really great) source of water dust
same with worms but for earth dust instead
the potion recipe system was no longer being used
there are many non-alch items that have a precursor attached to them,
useful for alchemists to be able to tell those on a glance
regarding the ozium thing, in my experience poppy plants give very few
poppies, so this should somewhat help in getting more ozium (which in my
experience was in short supply)

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Leeches and worms are grindable in a mortar
add: examining any item that has a precursor will tell you what essence
it gives (with enough alchemy skill of course)
del: Removed grind recipe for organs, grinding them gives their reagents
instead
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->