Palace changes (#3490)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
<img width="628" height="673" alt="image"
src="https://github.com/user-attachments/assets/fd548364-4cab-4389-8a21-3117d05f4e67"
/>
<img width="714" height="564" alt="image"
src="https://github.com/user-attachments/assets/aba4d555-bff2-4c88-a7d7-a62a6c11d620"
/>
<img width="625" height="550" alt="image"
src="https://github.com/user-attachments/assets/1242ae22-f283-40b2-a030-03cf2391008f"
/>

Shrinks the existing tavern along with a few minor fixes to the church.
Makes it more playable for single barkeep runs.
## Why It's Good For The Game

Requests made by players of the game, this was repeated over and over by
multiple people so we'll see how it goes.

fixes an oversized inn the people have had issues with, makes the
kitchen and bar closer to allow for ease of play.
Addition of the cannon to grenzelhoftian fort
Addition of another secret under the apothecary (possible fusion of
apothecary and medical soon.)
Fixes over the map in general.
## Changelog

:cl:
map: Updated rosewood.
map: large sections of the map to shrink it down to a more manageable
size.
map: Shrinked the lower floor of the inn.
add: Adds the sword display, paper and quills and another set of keys to
the inn.
del: The bathrooms and bath from the inn.
add: Cannon to the fort
add: A secret to the apothecary
/:cl:

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->