Remove stickybans (#3660)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Port https://github.com/Monkestation/Monkestation2.0/pull/8869

This PR removes stickybans in its entirety as nobody uses it, nobody
should be using it, and it is an awful system that can fuck up and ban
everyone with in a 1 morbillion mile radius if used incorrectly on
pre-builts.

<img width="669" height="397" alt="image"
src="https://github.com/user-attachments/assets/35170a0e-596c-4c30-a375-89efdeb820af"
/>


There are only 5 stickybans in the Vanderlin DB and they're from the DB
copy we did when Vanderlin was first created

## Why It's Good For The Game

Awful 
## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
del: Remove Stickybans
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->