Prayer and psycross changes (#3532)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

You can pray to Psydon and The Ten if you have a psycross in your active
hand, on your neck, or on your wrists
for psydon its any psycross (like it always was), for the ten you need
their respective psycross/amulet
changes the path of the steel malum amulet and the noc amulet, noc
amulet is now made of silver
Reduced the amount of hammering needed for the steel malum amulet

## Why It's Good For The Game

Because being unable to pray to your patron despite having an amulet
kind of sucked I think
Also the noc amulet was meant to be silver according to the comment so I
made it so

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
balance: You can pray to Psydon (and The Ten with their respective
amulet) if you have a psycross/amulet in your active hand, neck or
wrists
fix: Noc amulet is now made of silver as it shouldve been
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->