Airgun (#3049)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

I vaguely recall there was an idea to add an airgun of sorts
this PR is meant to be that

Somewhat similar to what the puffer (aka flintlock pistol) does, the gun
requires some set up before it can be fired
In regards to damage, it's both weaker and stronger. It can be fired to
deal 2/3rds damage of the lead ball, a normal amount, or 4/3rds. This
however consumes more steam (which needs to be recharged at a steam
injector mind you).
So I think it's balanced, by the fact that it's a bit unwieldy to use,
requires steam (and so it requires recharging) and requires the
knowledge on how to use it.
It's craftable 1 steel, 2 cured leather, 3 bronze and 4 metal gears.

<img width="456" height="177" alt="image"
src="https://github.com/user-attachments/assets/a685a890-5e4b-41fe-bbd7-8995e7bf1ca9"
/>
<img width="162" height="118" alt="image"
src="https://github.com/user-attachments/assets/69ee8dca-43da-4fb6-b74b-189b37520ec3"
/>

Sprite by terracotta

## Why It's Good For The Game

I think more steampunk weaponry is cool. The gun is expensive to make.
Both in terms of literal physical resources, but also time required to
get those resources, and the fact that you need MASTER engineering to
even have a reliable chance at making it.

## Changelog

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Adds the airgun, craftable by artificiers. Basically a steam
musket. Requires average engineering to even be able to use it.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->