Why are all these node tooltips so bad (#3674)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
seriously they're so bad read em
skill tooltips too
(also moves xylixes cruel prank such that if you zoom out *aaaaalll* the
way, you might glance it out of the corner of your eye. **funny**)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Informative tooltips not outright lying to you is good. (some of them
were literally just not what the spell did????)
moving xylixs to the point where you might be able to see it. this is
not good but it is funny and mostly harmless.
(if you click a button labeled "cruel prank" and "fucked up and evil"
that's on you if it plays a cruel prank on you that is fucked up and
evil methinks)
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl: notably magickal frog
spellcheck: The magician in charge of writing the spell tree was found
to be a Zizoite who was intentionally misinforming many novice wizards.
They have been sacked, socked, and subsequently shot. Hail the Ten.
spellcheck: Some skill descriptions are more accurate too.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->