fixes/changes to gem carving and two classes (#3737)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This pr is meant to fix some gem carving recipes, adjusts the lore for
the lakkarian cleric, and gives the ex-pegasus knight riding skill.

## Why It's Good For The Game
I noticed that one gem carving recipe was broken, so i needed to fix it,
I also made it so that the cut gem recipes require no skill to do, so
that way if a player stumbles upon a gem but has no crafting skill, they
can atleast carve one recipe. Also, I forgot to give ex-pegasus knights
riding skill, so this pr fixes that as well.
## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Added different lore for the lakkarian cleric
qol: gave ex-pegasus knight riding skill and made one gem carving recipe
require no skill
fix: fixed a broken recipe
:cl:
<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->