Cloth and Plain Robes Fix (#3724)
## About The Pull Request
Makes basic cloth robes craftable by giving them and the plain-dyed
version unique names in the sewing menu.

Currently, it will always attempt to craft the diff3 version regardless
of which you pick.
## Why It's Good For The Game
More crafting options and more UI clarity.
## Changelog
:cl:
fix: Changed the names of two instances of "robes" to be unique in the
sewing menu.
/:cl:
## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.