The Gun Buff... (#3429)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Puffer damage does not scale of PER anymore, instead it is a flat 130.

- The puffer can't delimb and has a reduced crit chance.

## Why It's Good For The Game

- I think most players would agree that the puffer is not strong
specially when people have experienced fighting the renegade, it is
truly a rare skill that few have it and the fewer can use the puffer
effectily since it scales of the user skill for aim, reload and etc

- Argueably a crossbow with enough skill and STR has a greater effect
than a single puffer user, this would make the weapon really dangerous
and i can make the renegade puffer weapon a weaker version.



## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
bal: Buff Puffer damage and it does not scale with PER anymore.
bal: Puffer can't delimb and has a reduced crit chance.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>