Brushwalking for Grn. Blessing, Poachers, and Ruffians (#3728)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds a "brushwalker" trait for the briar's green blessing. Gives it to
knave poachers and forest ruffians at roundstart. This would let them
move through meagre bushes and bog bushes, (but not thorn bushes,)
without immobilization or cuts.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

I was always confused as to why the green blessing granted webwalking
when Dendor has disowned the creatures of the caves. Now that bees and
apiaries are in the game they're a much more thematically-appropriate
source of honey, too, so this gives the blessing some situational
utility as opposed to basically never being relevant.

For bandits, it fits the poacher theme and helps incentivize picking
bows over crossbows. For ruffians, it fits one of their roles as the
scouts of the forest guard and their relatively low combat skills/stats
means it shouldn't be too abusable. I didn't want to _just_ give it to
briars and am of course open to more/fewer/different jobs having it if
there's any concern about balance.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Added a trait called "Brushwalker," which lets you freely pass
meagre bushes and bog bushes.
del: Removed the "Webwalker" trait from briar's grn. blessing and
reference to it from largely-unused tangler mob.
balance: Gave the "Brushwalker" trait to briar's grn. blessings, bandit
knave poachers, and forest ruffians.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->