Steam Armor Balance (#3726)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

First a nerf, by hitting the chest of a steam knight their boiler will
get damaged, it won't be the full damage but a fraction, but at the same
time the more damage the boiler takes the more steam it will drain and
eventually breaking it shutdowing the armor buff

Second a nerf and a buff, the helmet FOV is back to the old 70 while the
steam armor is off but when it is powered up it goes back to the old 180
but they also gain a partial night vision that helps them see at
night/dark.

## Why It's Good For The Game

Steam Knight is very strong so i hope this balances out, the numbers can
be tweaked for more or less

## Changelog

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
balance: The boiler can take damage by hitting the chest of a steam
knight.
balances The steamknight helmet fov is at 70 unpowered and 180 powered
while powered they also gain a partial night vision.
balance: The more damaged the boiler is the more steam it will drain.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>