Tainted Lux (Rousman, goblin, orc, zizombie lux) (#3582)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This needs to be TMed after the lux surgery PR is merged.

By being touched or created by the gods the races gained lux but not all
lux is the same, specially those who were touched by the inhumen such as
goblins, orcs, rousman and zizombies, those luxes are tainted.. very to
little value and even tho it can help fill some emptyiness of the soul,
it is not the same.

Tainted lux is harder to manipulate and upon infusing/reviving someone
with them there is a 50% chance of that person getting cursed with a -4
Fortune effect forever, not everyone can handle the tainted lux unless
you are molded with it.

Having tainted lux also gives the player -2 Fortune.

Tainted races such as Goblins, rousman and Orcs are not affected by the
malus of tainted lux but they can only be revived/bestowed tainted luxes
as well.

Tainted lux gif:

![tainted_lux](https://github.com/user-attachments/assets/3a02320d-eb6b-45f3-8255-414c148653b0)


## Why It's Good For The Game

Helps with the game being more lore accurate (feel free to correct me if
i am wrong), reduces the npcs lux farming yiedling basically infinite
mammons since each lux sells for 500 now unless you are removing it from
someone it will be basically worthless, those without lux have a cheaper
alternative but with downsides

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Tainted Lux
bal: Npcs such as : rousman, goblins, orcs and zizombies have tainted
lux instead of lux.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>